+++ etc/function-definitions.json
@@ -365,56 +365,56 @@
     },
     "fmax": {
         "sources": [
-            "src/math/fmax.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmax.rs"
         ],
         "type": "f64"
     },
     "fmaxf": {
         "sources": [
-            "src/math/fmaxf.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmax.rs"
         ],
         "type": "f32"
     },
     "fmaxf128": {
         "sources": [
-            "src/math/fmaxf128.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmax.rs"
         ],
         "type": "f128"
     },
     "fmaxf16": {
         "sources": [
-            "src/math/fmaxf16.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmax.rs"
         ],
         "type": "f16"
     },
     "fmin": {
         "sources": [
-            "src/math/fmin.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmin.rs"
         ],
         "type": "f64"
     },
     "fminf": {
         "sources": [
-            "src/math/fminf.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmin.rs"
         ],
         "type": "f32"
     },
     "fminf128": {
         "sources": [
-            "src/math/fminf128.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmin.rs"
         ],
         "type": "f128"
     },
     "fminf16": {
         "sources": [
-            "src/math/fminf16.rs",
+            "src/math/fmin_fmax.rs",
             "src/math/generic/fmin.rs"
         ],
         "type": "f16"
