+++ src/math/mod.rs
@@ -165,10 +165,7 @@ mod floor;
 mod floorf;
 mod fma;
 mod fmaf;
-mod fmax;
-mod fmaxf;
-mod fmin;
-mod fminf;
+mod fmin_fmax;
 mod fmod;
 mod fmodf;
 mod frexp;
@@ -273,10 +270,7 @@ pub use self::floor::floor;
 pub use self::floorf::floorf;
 pub use self::fma::fma;
 pub use self::fmaf::fmaf;
-pub use self::fmax::fmax;
-pub use self::fmaxf::fmaxf;
-pub use self::fmin::fmin;
-pub use self::fminf::fminf;
+pub use self::fmin_fmax::{fmax, fmaxf, fmin, fminf};
 pub use self::fmod::fmod;
 pub use self::fmodf::fmodf;
 pub use self::frexp::frexp;
@@ -346,8 +340,6 @@ cfg_if! {
         mod fabsf16;
         mod fdimf16;
         mod floorf16;
-        mod fmaxf16;
-        mod fminf16;
         mod fmodf16;
         mod ldexpf16;
         mod rintf16;
@@ -363,8 +355,8 @@ cfg_if! {
         pub use self::fabsf16::fabsf16;
         pub use self::fdimf16::fdimf16;
         pub use self::floorf16::floorf16;
-        pub use self::fmaxf16::fmaxf16;
-        pub use self::fminf16::fminf16;
+        pub use self::fmin_fmax::fmaxf16;
+        pub use self::fmin_fmax::fminf16;
         pub use self::fmodf16::fmodf16;
         pub use self::ldexpf16::ldexpf16;
         pub use self::rintf16::rintf16;
@@ -385,8 +377,6 @@ cfg_if! {
         mod fdimf128;
         mod floorf128;
         mod fmaf128;
-        mod fmaxf128;
-        mod fminf128;
         mod fmodf128;
         mod ldexpf128;
         mod rintf128;
@@ -403,8 +393,8 @@ cfg_if! {
         pub use self::fdimf128::fdimf128;
         pub use self::floorf128::floorf128;
         pub use self::fmaf128::fmaf128;
-        pub use self::fmaxf128::fmaxf128;
-        pub use self::fminf128::fminf128;
+        pub use self::fmin_fmax::fmaxf128;
+        pub use self::fmin_fmax::fminf128;
         pub use self::fmodf128::fmodf128;
         pub use self::ldexpf128::ldexpf128;
         pub use self::rintf128::rintf128;
